refactor(asset): merge asset-group into asset/types
